# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic REST API using Golang",
        "title": "Example Go REST API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get All Products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetProductsResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create New Product",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CreateProductResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get Products by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetProductByIdResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UpdateProductResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ProductResponse"
                },
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "GetProductByIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ProductResponse"
                },
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "GetProductsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ProductResponse"
                    }
                },
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "ProductRequest": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "ProductResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "UpdateProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ProductResponse"
                },
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        }
    }
}